+++ samplesheets/src/components/renderers/DataCellRenderer.vue
@@ -141,7 +141,7 @@ export default Vue.extend(
           let contactGroup = contactRegex.exec(this.value)
           return {name: contactGroup[1], email: contactGroup[2]}
         } else {
-          this.colType = null  // Fall back to standard field
+          this.colType = null // Fall back to standard field
         }
       },